requirements.txt
accelerate==0.33.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
attrs==25.4.0
audioread==3.1.0
bcrypt==4.0.1
bitsandbytes==0.48.2
boto3==1.40.67
botocore==1.40.67
brotli==1.2.0
cached_path==1.8.0
cachetools==6.2.1
celluloid==0.2.0
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
click==8.1.8
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
datasets==4.4.1
decorator==5.2.1
deepspeed==0.18.2
dill==0.4.0
ecdsa==0.19.1
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
f5-tts==1.1.9
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.19.1
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.4
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.45.1
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
ninja==1.13.0
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-whisper==20250625
orjson==3.11.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
plac==1.4.5
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.9
ptflops==0.7.5
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
-e git+https://github.com/resemble-ai/resemble-enhance.git@8e978149bfe8abab3eb77d965d579a111afdb0ff#egg=resemble_enhance
rich==13.9.4
rsa==4.9.1
ruff==0.14.3
s3transfer==0.14.0
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soxr==1.0.0
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.1
torchdiffeq==0.2.5
torchvision==0.24.0
tqdm==4.67.1
transformers==4.35.2
transformers-stream-generator==0.0.5
triton==3.5.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
unidic==1.1.0
urllib3==1.26.13
uvicorn==0.24.0
uvloop==0.22.1
vocos==0.1.0
wandb==0.22.3
wasabi==0.10.1
watchfiles==1.1.1
websockets==15.0.1
x-transformers==2.11.15
xxhash==3.6.0
yarl==1.22.0
